// src/types/profile.ts
export interface ProfileData {
  name?: string;
  age?: number;
  job?: string;
  hobby?: string;
  residence?: string;
  workplace?: string;
  bloodType?: string;
  education?: string;
  workType?: string;
  holiday?: string;
  marriageHistory?: string;
  hasChildren?: string;
  smoking?: string;
  drinking?: string;
  livingWith?: string;
  marriageIntention?: string;
  selfIntroduction?: string;
}